example/generate_lookups/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
importlib-metadata==4.8.3
mysql-connector-python==8.0.29
packaging==21.3
protobuf==3.19.4
pyparsing==3.0.8
redis==4.3.1
typing_extensions==4.1.1
wrapt==1.14.1
zipp==3.6.0
